feat: updated file name to usePublishQueue
